allow options for buy method
account_id 
Specify which account is used for crediting bought amount. The default is your primary account.
Value: Must be String

agree_btc_amount_varies 
Whether or not you would still like to buy if you have to wait for your money to arrive to lock in a price.
Value: Must be Boolean

payment_method_id 
The ID of the payment method that should be used for the buy. Payment methods can be listed using the /payment_methods API call.
Value: Must be String